feat: finish

+++ src/routes/user.ts
@@ -7,6 +7,15 @@ const userService = new UserService();
 
 export const userRouter = server.router(userContract, {
   async createUser(input) {
+    const existingUser = await userService.findByEmail(input.body.email);
+    if (existingUser) {
+      return {
+        status: 400,
+        body: {
+          message: 'User already exists',
+        },
+      };
+    }
     const user = await userService.create(input.body);
     if (!user) {
       return {
@@ -30,6 +39,15 @@ export const userRouter = server.router(userContract, {
     };
   },
   async deleteUser(input) {
+    const user = await userService.show(Number(input.params.id));
+    if (!user) {
+      return {
+        status: 404,
+        body: {
+          message: 'User not found',
+        },
+      };
+    }
     await userService.destroy(Number(input.params.id));
     return {
       status: 200,
@@ -39,6 +57,15 @@ export const userRouter = server.router(userContract, {
     };
   },
   async updateUser(input) {
+    const userExists = await userService.show(Number(input.params.id));
+    if (!userExists) {
+      return {
+        status: 404,
+        body: {
+          message: 'User not found',
+        },
+      };
+    }
     const user = await userService.update(Number(input.params.id), input.body);
     if (!user) {
       return {
